chore: Convert sheets on `/setting/scanning` route to True Sheet
- We encountered some broken behavior during this move (ie: input focus/focusing behavior).
  - For some reason, we can focus fine on the text inputs in `<ScanFilterListSheet />` but not the numeric input for `<MinDurationSheet />`.
- We set the keyboard mode to `pan` for the `<ScanFilterListSheet />` as we don't want the sheet to grow.
  - This has the downside of not allowing us to scroll through all the available items when the keyboard is open.
- For the swipeables to work inside the True Sheet, we need a new `GestureHandlerRootView` in it.
