+++ mobile/src/screens/Sheets/Settings/helpers/ScanFilterData.ts
